backendtldr/summarize/requirements.txt
django-tagulous==0.12.0
nltk==3.2.4
tensorflow==1.2.1

